:zap: improve bottom navigation and folder structure
