+++ app/src/main/java/com/aditya/kitchengenius/presentation/activity/MainActivity.kt
@@ -1,0 +1,22 @@
+package com.aditya.kitchengenius.presentation.activity
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import androidx.navigation.findNavController
+import androidx.navigation.ui.setupWithNavController
+import com.aditya.kitchengenius.R
+import com.google.android.material.bottomnavigation.BottomNavigationView
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        val splashScreen = installSplashScreen()
+
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottomMenuNav)
+        val navController = findNavController(R.id.navHostFragment)
+        bottomNavigationView.setupWithNavController(navController)
+    }
+}
