+++ app/src/main/java/com/aditya/kitchengenius/presentation/fragment/favorite/FavoriteFragment.kt
@@ -1,0 +1,23 @@
+package com.aditya.kitchengenius.presentation.fragment.favorite
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.aditya.kitchengenius.R
+
+
+class FavoriteFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_favorite, container, false)
+    }
+}
