+++ app/src/main/java/com/aditya/kitchengenius/presentation/fragment/home/HomeFragment.kt
@@ -1,0 +1,43 @@
+package com.aditya.kitchengenius.presentation.fragment.home
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
+import com.aditya.kitchengenius.databinding.FragmentHomeBinding
+
+
+class HomeFragment : Fragment() {
+    private var _binding: FragmentHomeBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    private fun hideToolbar() {
+        val activity = requireActivity() as AppCompatActivity
+        activity.supportActionBar?.hide()
+    }
+
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        hideToolbar()
+    }
+
+
+
+}
