+++ app/src/main/java/com/aditya/kitchengenius/presentation/fragment/profile/ProfileFragment.kt
@@ -1,17 +1,15 @@
-package com.aditya.kitchengenius
+package com.aditya.kitchengenius.presentation.fragment.profile
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.aditya.kitchengenius.databinding.FragmentHomeBinding
+import com.aditya.kitchengenius.R
 
 
-class HomeFragment : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
-    private val binding get() = _binding!!
 
+class ProfileFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
     }
@@ -19,11 +17,9 @@ class HomeFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
-        return binding.root
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_profile, container, false)
     }
-
-
 
 }
