+++ app/src/main/res/layout/fragment_onboarding.xml
@@ -4,9 +4,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".OnboardingFragment">
+    tools:context=".presentation.fragment.onboarding.OnboardingFragment">
 
     <ImageView
+        android:contentDescription="@string/onboarding_image"
         android:scaleType="centerCrop"
         android:src="@drawable/onboarding_image"
         android:layout_width="match_parent"
@@ -18,6 +19,7 @@
         android:text="@string/hello_blank_fragment" />
 
     <ImageView
+        android:contentDescription="@string/icon_kitchen_genius"
         android:id="@+id/kitchenGeniusImageView"
         android:src="@drawable/ic_launcher_foreground_red"
         android:layout_width="200dp"
