+++ app/src/main/res/values-night/themes.xml
@@ -31,4 +31,15 @@
         <item name="android:fontFamily">@font/poppins_regular</item>
         <item name="android:textColor">@color/black</item>
     </style>
+
+    <style name="CustomTextInputLayout" parent="Widget.MaterialComponents.TextInputLayout.OutlinedBox">
+        <item name="hintTextColor">@color/hint_text_color_dark</item>
+        <item name="boxStrokeColor">@color/stroke_color_dark</item>
+        <item name="boxBackgroundColor">@color/background_color_dark</item>
+        <item name="boxCornerRadiusTopStart">8dp</item>
+        <item name="boxCornerRadiusTopEnd">8dp</item>
+        <item name="boxCornerRadiusBottomStart">8dp</item>
+        <item name="boxCornerRadiusBottomEnd">8dp</item>
+        <item name="android:textColorHint">@color/hint_text_color_dark</item>
+    </style>
 </resources>
